requirements.txt
fastapi~=0.104.1
uvicorn[standard]~=0.24.0
qdrant-client~=1.7.0
google-generativeai~=0.3.0
httpx~=0.23.0
pydantic~=2.5.0
python-dotenv~=1.0.0
